refactor: mark numba._soft_imports as private
