+++ skchange/utils/numba/__init__.py
@@ -1,3 +1,3 @@
 """Numba utility functions."""
 
-from .soft_import import jit, njit, numba_available, prange
+from ._soft_import import jit, njit, numba_available, prange
